refactor(gc): extract GC agent tests
